refactor: move match_node impl out
